cis-api/requirements.txt
aiofiles==0.5.0
alembic==1.7.0
aniso8601==7.0.0
anyio==3.3.4
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.24.0
asynctest==0.13.0
attrs==21.2.0
boto3==1.20.7
botocore==1.23.7
CacheControl==0.12.10
cached-property==1.5.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==7.1.2
cryptography==35.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
et-xmlfile==1.1.0
fastapi==0.69.0
FastAPI-SQLAlchemy==0.2.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.7
httptools==0.2.0
httpx==0.18.2
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
inflect==5.3.0
iniconfig==1.1.1
iso3166==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
keycloak==3.1.0
loguru==0.5.3
Mako==1.1.5
mapbox==0.18.0
MarkupSafe==2.0.1
msgpack==1.0.2
openpyxl==3.0.9
packaging==21.0
pendulum==2.1.2
pluggy==1.0.0
polyline==1.4.0
promise==2.3
psutil==5.8.0
psycopg2==2.9.2
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pydash==5.1.0
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.15.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
requests==2.26.0
resolvelib==0.5.4
rfc3986==1.5.0
rsa==4.7.2
Rx==1.6.1
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.23
sqlalchemy-json==0.4.0
sqlparse==0.4.2
starlette==0.15.0
toml==0.10.2
typing-extensions==3.10.0.2
ujson==4.1.0
uritemplate==4.1.1
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
Werkzeug==2.0.2
zipp==3.5.0
